cmd/go-clang-gen: move core function internally to make it testable
